Improved Discord module
