+++ src/discord/decorators/interfaces/ondecorator.options.ts
