--- src/discord/discord.config.ts
@@ -1,8 +1,0 @@
-import {ConfigModule, registerAs} from '@nestjs/config';
-import {DiscordModuleOptions} from './foundation/interfaces/discordmodule.options';
-
-export const DiscordConfig = ConfigModule.forFeature(
-    registerAs('discord', (): DiscordModuleOptions => ({
-        token: process.env.DISCORD_TOKEN,
-    })),
-);
