--- src/discord/foundation/interfaces/discordoptions.factory.ts
@@ -1,5 +1,0 @@
-import {DiscordModuleOptions} from './discordmodule.options';
-
-export interface DiscordOptionsFactory {
-    createDiscordOptions(): Promise<DiscordModuleOptions> | DiscordModuleOptions;
-}
