+++ src/discord/helpers/metadata.accessor.ts
