+++ src/discord/interfaces/discordmodule.options.ts
